+++ autoprecompiles/src/optimizer.rs
@@ -1,0 +1,182 @@
+use itertools::Itertools;
+use powdr_ast::analyzed::{AlgebraicReference, PolyID, PolynomialType};
+use powdr_constraint_solver::{
+    constraint_system::{BusInteraction, ConstraintSystem},
+    quadratic_symbolic_expression::QuadraticSymbolicExpression,
+    solver::Solver,
+    symbolic_expression::SymbolicExpression,
+};
+use powdr_number::FieldElement;
+use powdr_pilopt::{
+    qse_opt::{
+        algebraic_to_quadratic_symbolic_expression, quadratic_symbolic_expression_to_algebraic,
+        Variable,
+    },
+    simplify_expression,
+};
+
+use crate::{BusInteractionKind, SymbolicBusInteraction, SymbolicConstraint, SymbolicMachine};
+
+/// Simplifies the constraints as much as possible.
+/// This function is similar to powdr_pilopt::qse_opt::run_qse_optimization, except it:
+/// - Runs on the entire constraint system, including bus interactions.
+/// - Panics if the solver fails.
+/// - Removes trivial constraints (e.g. `0 = 0` or bus interaction with multiplicity `0`)
+///   from the constraint system.
+/// - Calls `simplify_expression()` on the resulting expressions.
+pub fn optimize<P: FieldElement>(symbolic_machine: SymbolicMachine<P>) -> SymbolicMachine<P> {
+    let constraint_system = symbolic_machine_to_constraint_system(symbolic_machine);
+
+    log_constraint_system_stats("Starting optimize()", &constraint_system);
+    let constraint_system = solver_based_optimization(constraint_system);
+    log_constraint_system_stats("After solver-based optimization", &constraint_system);
+    let constraint_system = remove_trivial_constraints(constraint_system);
+    log_constraint_system_stats("After removing trivial constraints", &constraint_system);
+
+    // TODO: Add equivalent of replace_linear_witness_columns step to make
+    // powdr_optimize_legacy obsolete
+
+    constraint_system_to_symbolic_machine(constraint_system)
+}
+
+fn symbolic_machine_to_constraint_system<P: FieldElement>(
+    symbolic_machine: SymbolicMachine<P>,
+) -> ConstraintSystem<P, Variable> {
+    ConstraintSystem {
+        algebraic_constraints: symbolic_machine
+            .constraints
+            .iter()
+            .map(|constraint| algebraic_to_quadratic_symbolic_expression(&constraint.expr))
+            .collect(),
+        bus_interactions: symbolic_machine
+            .bus_interactions
+            .iter()
+            .map(symbolic_bus_interaction_to_bus_interaction)
+            .collect(),
+    }
+}
+
+fn constraint_system_to_symbolic_machine<P: FieldElement>(
+    constraint_system: ConstraintSystem<P, Variable>,
+) -> SymbolicMachine<P> {
+    SymbolicMachine {
+        constraints: constraint_system
+            .algebraic_constraints
+            .iter()
+            .map(|constraint| SymbolicConstraint {
+                expr: simplify_expression(quadratic_symbolic_expression_to_algebraic(constraint)),
+            })
+            .collect(),
+        bus_interactions: constraint_system
+            .bus_interactions
+            .into_iter()
+            .map(bus_interaction_to_symbolic_bus_interaction)
+            .collect(),
+    }
+}
+
+fn solver_based_optimization<T: FieldElement>(
+    constraint_system: ConstraintSystem<T, Variable>,
+) -> ConstraintSystem<T, Variable> {
+    let result = Solver::new(constraint_system)
+        .solve()
+        .map_err(|e| {
+            panic!("Solver failed: {e:?}");
+        })
+        .unwrap();
+    log::trace!("Solver figured out the following assignments:");
+    for (var, value) in result.assignments.iter() {
+        log::trace!("  {var} = {value}");
+    }
+    result.simplified_constraint_system
+}
+
+fn remove_trivial_constraints<P: FieldElement>(
+    mut symbolic_machine: ConstraintSystem<P, Variable>,
+) -> ConstraintSystem<P, Variable> {
+    let zero = QuadraticSymbolicExpression::from(P::zero());
+    symbolic_machine
+        .algebraic_constraints
+        .retain(|constraint| constraint != &zero);
+    symbolic_machine
+        .bus_interactions
+        .retain(|bus_interaction| bus_interaction.multiplicity != zero);
+    symbolic_machine
+}
+
+fn symbolic_bus_interaction_to_bus_interaction<P: FieldElement>(
+    bus_interaction: &SymbolicBusInteraction<P>,
+) -> BusInteraction<QuadraticSymbolicExpression<P, Variable>> {
+    BusInteraction {
+        bus_id: SymbolicExpression::Concrete(P::from(bus_interaction.id)).into(),
+        payload: bus_interaction
+            .args
+            .iter()
+            .map(|arg| algebraic_to_quadratic_symbolic_expression(arg))
+            .collect(),
+        multiplicity: algebraic_to_quadratic_symbolic_expression(&bus_interaction.mult),
+    }
+}
+
+fn bus_interaction_to_symbolic_bus_interaction<P: FieldElement>(
+    bus_interaction: BusInteraction<QuadraticSymbolicExpression<P, Variable>>,
+) -> SymbolicBusInteraction<P> {
+    // We set the bus_id to a constant in `bus_interaction_to_symbolic_bus_interaction`,
+    // so this should always succeed.
+    let id = bus_interaction
+        .bus_id
+        .try_to_number()
+        .unwrap()
+        .to_arbitrary_integer()
+        .try_into()
+        .unwrap();
+    SymbolicBusInteraction {
+        id,
+        // TODO: The kind of SymbolicBusInteraction is ignored, this field should be removed
+        kind: BusInteractionKind::Send,
+        args: bus_interaction
+            .payload
+            .into_iter()
+            .map(|arg| simplify_expression(quadratic_symbolic_expression_to_algebraic(&arg)))
+            .collect(),
+        mult: simplify_expression(quadratic_symbolic_expression_to_algebraic(
+            &bus_interaction.multiplicity,
+        )),
+    }
+}
+
+fn log_constraint_system_stats<P: FieldElement>(
+    step: &str,
+    constraint_system: &ConstraintSystem<P, Variable>,
+) {
+    let num_constraints = constraint_system.algebraic_constraints.len();
+    let num_bus_interactions = constraint_system.bus_interactions.len();
+    let num_witness_columns = constraint_system
+        .algebraic_constraints
+        .iter()
+        .flat_map(|constraint| constraint.referenced_variables())
+        .chain(
+            constraint_system
+                .bus_interactions
+                .iter()
+                .flat_map(|bus_interaction| bus_interaction.referenced_variables()),
+        )
+        .filter_map(|expr| {
+            if let Variable::Reference(AlgebraicReference {
+                poly_id:
+                    PolyID {
+                        ptype: PolynomialType::Committed,
+                        id,
+                    },
+                ..
+            }) = expr
+            {
+                Some(id)
+            } else {
+                None
+            }
+        })
+        .unique()
+        .count();
+    log::info!("{step} - Constraints: {num_constraints}, Bus Interactions: {num_bus_interactions}, Witness Columns: {num_witness_columns}");
+}
